refactor : errorMessage class name  and channel method name change
