+++ codeit-bootcamp-spring/1-sprint-mission/src/main/java/com/sprint/mission/discodeit/common/error/ErrorMessage.java
@@ -1,6 +1,6 @@
-package com.sprint.mission.discodeit.common.error.user;
+package com.sprint.mission.discodeit.common.error;
 
-public enum UserErrorMessage {
+public enum ErrorMessage {
 
     /**
      * 1000 ~ 1999 User 관련 에러
@@ -10,13 +10,15 @@ public enum UserErrorMessage {
     USER_NAME_NULL(1001, "유저의 이름은 반드시 존재해야합니다."),
 
     NAME_LENGTH_ERROR_MESSAGE(1002, "유저 이름의 길이 제한을 확인해주세요."),
+
+    USER_NOT_PARTICIPATED_CHANNEL(1010, "유저가 참여하지 않은 방입니다."),
     ;
 
     private final int errorCode;
     private final String message;
 
 
-    UserErrorMessage(int errorCode, String message) {
+    ErrorMessage(int errorCode, String message) {
         this.errorCode = errorCode;
         this.message = message;
     }
