+++ codeit-bootcamp-spring/1-sprint-mission/src/main/java/com/sprint/mission/discodeit/common/error/user/UserException.java
@@ -1,5 +1,7 @@
 package com.sprint.mission.discodeit.common.error.user;
 
+import com.sprint.mission.discodeit.common.error.ErrorMessage;
+
 public class UserException extends RuntimeException {
 
     public UserException(String message) {
@@ -10,8 +12,17 @@ public class UserException extends RuntimeException {
         super(message, cause);
     }
 
-    public static UserException of(UserErrorMessage message) {
+    public static UserException of(ErrorMessage message) {
         return new UserException(message.getMessage());
+    }
+
+    public static UserException errorMessageAndId(ErrorMessage message, String id) {
+        var format = String.format("%s : not participated channel id %s", message.getMessage(), id);
+        return new UserException(format);
+    }
+    public static UserException errorMessageAndChannelName(ErrorMessage message, String channelName) {
+        var format = String.format("%s : not participated channel name %s", message.getMessage(), channelName);
+        return new UserException(format);
     }
 
     public static UserException of(String message, Throwable cause) {
